% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_Z_OT.R
\name{approximate_strauss_constant_OT}
\alias{approximate_strauss_constant_OT}
\title{approximate Strauss constant using Ogata&Tanemura approx.}
\usage{
approximate_strauss_constant_OT(beta, gamma, range, bbox, Nmax = 10000, n,
  deg = 2, ...)
}
\description{
approximate Strauss constant using Ogata&Tanemura approx.
}

