% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_distance.R
\name{bbox_distance}
\alias{bbox_distance}
\title{distance from points to bbox border}
\usage{
bbox_distance(x, bbox)
}
\arguments{
\item{x}{coordinate matrix}

\item{bbox}{bounding box}
}
\description{
distance from points to bbox border
}

