% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fstrauss.direct}
\alias{fstrauss.direct}
\title{Fit Strauss process to data using maximum approx. likelihood}
\usage{
fstrauss.direct(x, R, lower = c(1e-09, 1e-09), upper = c(Inf, 1), init, ...)
}
\arguments{
\item{x}{a list with elements $x:data.frame of locations (ncol 2 or 3), $bbox: matrix with columns giving window bounds}

\item{R}{the fixed known interaction range. Use profile pl to infer this.}

\item{lower}{Lower limits for searching (beta,gamma).}

\item{upper}{Upper limits for searching (beta,gamma).}

\item{init}{Initial values for optim.}

\item{...}{passed on to \code{\link{approximate_strauss_constant}}

Reduced sample border correction with radius R.

See \link{approximate_strauss_constant} for approximation options.}
}
\description{
ML estimate of beta and gamma using optimization of log-likelihood,
where the normalizing constant is approximated directly.
}

