% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_vbpl.R
\name{fstrauss.vb}
\alias{fstrauss.vb}
\title{Mean field fit of Strauss process using pseudolikelihood approximation}
\usage{
fstrauss.vb(x, range, nx = 25, eps = 1e-04, maxiter = 100, verb = FALSE)
}
\description{
Not yet implemented properly.
}

