% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_pots.R
\name{gibbs_pot}
\alias{gibbs_pot}
\title{Computes the sufficient statistics for exp-Gibbs models}
\usage{
gibbs_pot(x, par, model, ...)
}
\description{
Computes the sufficient statistics for exp-Gibbs models
}
