% Generated by roxygen2 (4.0.2): do not edit by hand
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Compute the marginal likelihood for Bayesian Strauss inference}
\usage{
marginal_likelihood(stats, R, bbox, priors, mciter = 1000)
}
\description{
Compute the marginal likelihood for Bayesian Strauss inference
}

