% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{strauss_beta}
\alias{strauss_beta}
\title{Solve beta for Strauss given on n,gamma,range}
\usage{
strauss_beta(n, gamma, R, dim = 3, stucki = FALSE, take = mean)
}
\arguments{
\item{n}{number of points}

\item{gamma}{gamma parameter}

\item{R}{range}

\item{dim}{dimension}

\item{stucki}{see details}

\item{take}{see details}
}
\description{
Solve beta for Strauss given on n,gamma,range
}
\details{
If stucki=FALSE, use Poisson mgf approximation, 
else use the function 'take' (default mean) on the bounds given by Stucki 2012.
'take' could be mean, max or min, for example.
see \link{strauss_intensity}
}

