% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{strauss_intensity}
\alias{strauss_intensity}
\title{Approximate intensity for Strauss based on beta,gamma,range}
\usage{
strauss_intensity(beta, gamma, range, dim = 3, stucki = FALSE)
}
\arguments{
\item{beta}{beta parameter}

\item{gamma}{gamma parameter}

\item{dim}{dimension}

\item{stucki}{see details}

\item{R}{range}
}
\description{
Approximate intensity for Strauss based on beta,gamma,range
}
\details{
if stucki=FALSE, use Poisson mgf approximation, 
else return the limits given in Stucki 2012 paper, which are 
accurate but bounds. The average is quite good.
}
