% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{strauss_pcf}
\alias{strauss_pcf}
\title{Pair correlation bounds given beta, gamma, R, lambda}
\usage{
strauss_pcf(r, beta, gamma, R, lambda, dim = 2)
}
\arguments{
\item{r}{range vector}

\item{beta}{beta parameter}

\item{gamma}{gamma parameter}

\item{R}{interaction range}

\item{lambda}{intensity}

\item{dim}{dimension}
}
\description{
Pair correlation bounds given beta, gamma, R, lambda
}
\details{
Use \code{\link{strauss_intensity}} to get a 'lambda' (approximation) for input.

The function returns the upper and lower bounds derived by Stucki & Schuhmacher 2013. The bounds are good for small beta.
}
