% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{strauss_theory}
\alias{strauss_theory}
\title{Theoretical maximum for the range parameter for hard core}
\usage{
strauss_theory(intensity, dim = 3)
}
\description{
Given the number of points (intensity) and the gamma=0, what is the 
theoretical upper limit for range?
}
