% Generated by roxygen2 (4.0.2): do not edit by hand
\name{approximate_strauss_constant}
\alias{approximate_strauss_constant}
\title{Approximate Strauss constant (master function)}
\usage{
approximate_strauss_constant(beta, gamma, range, bbox, method = "R", ...)
}
\arguments{
\item{beta}{beta}

\item{gamma}{gamma}

\item{range}{range}

\item{bbox}{bounding box}

\item{method}{OT, PS, P, R or R3. Default: R}

\item{...}{further details for methods

For all methods but "PS" conditioning on number of points is by setting the named parameter n.

Methods:

OT : Direct approx., Ogata&Tanemura 1981. \code{\link{approximate_strauss_constant_OT}}

P: Direct approx., Penttinen 1998. \code{\link{approximate_strauss_constant_penttinen}}

PS: MC unbiased estimate, Path sampling Berthelsen&Moller 2003. \code{\link{approximate_strauss_constant_PS}}

R: My own version of Penttinen for unconditional case. Based on a normal approximation (not published). Works when beta*Area >> 0

R3: Third order approx. Same as R but with extra term. (See Ripley 88 p.60-62).\code{\link{approximate_strauss_constant_R3}}

The returned value is in log-scale.}
}
\description{
Approximate the (log of) normalizing constant in the Strauss density. Log-scale.
}
\examples{
beta <- 100
gamma <- 0.1
R <- 0.05
bbox <- cbind(0:1, 0:1)
for(m in c("P", "PS", "R3")) print(approximate_strauss_constant(beta, gamma, R, bbox, m))

approximate_strauss_constant(beta, gamma, R, bbox, n=100)
}

