\name{approximate_strauss_constant_penttinen}
\alias{approximate_strauss_constant_penttinen}
\title{approximate Strauss constant using Penttinen84 approx.}
\usage{
approximate_strauss_constant_penttinen(beta, gamma, range, bbox, Nmax = 10000)
}
\description{
approximate Strauss constant using Penttinen84 approx.
}

