% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_Z_R.R
\name{approximate_strauss_constant_R}
\alias{approximate_strauss_constant_R}
\title{approximate unconditional Strauss constant using Gaussian approx. of the expectation}
\usage{
approximate_strauss_constant_R(beta, gamma, range, bbox, n, ...)
}
\description{
The same Poisson approximation for Z_n as in Penttinen, 
but the unconditional is via Gaussian expectation
instead of numerical integration.
}
\details{
See \link{approximate_strauss_constant_penttinen} .
}
