% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dstrauss}
\alias{dstrauss}
\title{density of a strauss process}
\usage{
dstrauss(x, beta, gamma, range, method = "OT", ...)
}
\arguments{
\item{x}{list with x: coordinate matrix (col-wise), bbox: bounding box (col-wise ranges)}

\item{beta}{beta parameter}

\item{gamma}{gamma parameter}

\item{range}{interaction range}

\item{method}{how to approximate the constant.}

\item{...}{passed to constant approximator.

See \code{\link{approx_strauss_constant}} for the methods

The value is in logarithm scale}
}
\description{
Now only for cuboidal window
}

