% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood.R
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Compute the marginal likelihood for Bayesian Strauss inference}
\usage{
marginal_likelihood(stats, R, bbox, priors, mciter = 1000)
}
\description{
Compute the marginal likelihood for Bayesian Strauss inference
}
